+++ CHANGELOG.md
@@ -5,14 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.2.1 - 2018-08-22
+## [0.3.0] - Unreleased
+
+### Added
+
+- [#3] `continueOnError()` option
+
+### Removed
+
+- [#3] `defaultHandler()` option
+
+## [0.2.1] - 2018-08-22
 
 ### Changed
 
 - Empty paths are always returned as `/`.
 - A `/` prefix works properly as well.
 
-## 0.2.0 - 2018-08-22
+## [0.2.0] - 2018-08-22
 
 ### Changed
 
@@ -21,3 +31,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2018-08-22
 
 First version
+
+[#3]: https://github.com/middlewares/base-path-router/issues/3
+
+[0.3.0]: https://github.com/middlewares/base-path-router/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/middlewares/base-path-router/compare/v0.2.0...v0.2.1
+[0.2.0]: https://github.com/middlewares/base-path-router/compare/v0.1.0...v0.2.0